InputSelectNew -> InputSelect
