+++ app/components/InputMemberPermissionSelect.tsx
@@ -2,11 +2,11 @@ import * as React from "react";
 import { useTranslation } from "react-i18next";
 import styled from "styled-components";
 import { s } from "@shared/styles";
-import { InputSelectNew, Option } from "~/components/InputSelectNew";
+import { InputSelect, Option } from "~/components/InputSelect";
 import { EmptySelectValue, Permission } from "~/types";
 
 type Props = Pick<
-  React.ComponentProps<typeof InputSelectNew>,
+  React.ComponentProps<typeof InputSelect>,
   "value" | "onChange" | "disabled" | "className"
 >;
 
@@ -45,6 +45,6 @@ export default function InputMemberPermissionSelect(
   );
 }
 
-const Select = styled(InputSelectNew)`
+const Select = styled(InputSelect)`
   color: ${s("textSecondary")};
 `;
