+++ app/scenes/Settings/Preferences.tsx
@@ -8,7 +8,7 @@ import { TeamPreference, UserPreference } from "@shared/types";
 import { Theme } from "~/stores/UiStore";
 import Button from "~/components/Button";
 import Heading from "~/components/Heading";
-import { InputSelectNew, Option } from "~/components/InputSelectNew";
+import { InputSelect, Option } from "~/components/InputSelect";
 import Scene from "~/components/Scene";
 import Switch from "~/components/Switch";
 import Text from "~/components/Text";
@@ -145,7 +145,7 @@ function Preferences() {
           </>
         }
       >
-        <InputSelectNew
+        <InputSelect
           options={languageOptions}
           value={user.language}
           onChange={handleLanguageChange}
@@ -159,7 +159,7 @@ function Preferences() {
         label={t("Appearance")}
         description={t("Choose your preferred interface color scheme.")}
       >
-        <InputSelectNew
+        <InputSelect
           options={themeOptions}
           value={ui.theme}
           onChange={handleThemeChange}
